requirements.txt
asgiref==3.2.10
astroid==2.4.2
Django==3.1
django-cors-headers==3.4.0
django-filter==2.3.0
djangorestframework==3.11.1
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
Pillow==7.2.0
pylint==2.5.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
whitenoise==5.2.0
wrapt==1.12.1
